inference/requirements.txt
confluent-kafka==2.2.0
asyncpg==0.29.0
PyYAML==6.0.1
opencv-python-headless==4.9.0.80
numpy==1.26.4